+++ .eslintrc.json
@@ -23,6 +23,8 @@
     "no-unused-vars": 1,
     "react/no-array-index-key": 0,
     "import/prefer-default-export": "off",
-    "react/jsx-props-no-spreading": "off"
+    "react/jsx-props-no-spreading": "off",
+    "react/jsx-uses-react": "off",
+    "react/react-in-jsx-scope": "off"
   }
 }
