+++ src/components/UI/LoadingSpinner.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import classes from './LoadingSpinner.module.css';
 
 const LoadingSpinner = () => (
