+++ src/components/doctor/DoctorPatients.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import { StyledPatientsList } from './DoctorPatients.styled';
 import EmptyList from './EmptyList';
