+++ src/components/doctor/Patient.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import {
   StyledAvatar, StyledDescription,
   StyledFooter,
