+++ src/components/forms/RestoreForm.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import * as Yup from 'yup';
 import { Formik } from 'formik';
 import Header from './header/header';
