+++ src/components/forms/SignUpForm.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import { Formik } from 'formik';
 import * as Yup from 'yup';
 import Header from './header/header';
