+++ src/components/forms/button/Button.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 import Arrow from '../../../assets/arrow-right.svg';
