+++ src/components/forms/footer/footer.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { StyledFooter, StyledLink, StyledText } from './footer.styled';
 
 const Footer = ({ text, path, link }) => (
