+++ src/components/forms/header/header.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useLocation } from 'react-router-dom';
 
 import { StyledArrow, StyledHeader, StyledText } from './header.styled';
