+++ src/components/user/createAppointment/checkboxes/Checkboxes.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import Checkbox from './Checkbox';
 
 const Checkboxes = ({ onChangeTime }) => {
