+++ src/pages/NotFound.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const NotFound = () => (
   <div className="centered">
     <p>Page not found!</p>
