+++ src/pages/UserCreateAppointment.js
@@ -1,5 +1,4 @@
-import React from 'react';
-import UserLayout from '../components/layout/UserLayout';
+import UserLayout from '../components/layout/User/UserLayout';
 import CreateAppointment from '../components/user/createAppointment/CreateAppointment';
 
 const UserCreateAppointment = () => (
