+++ src/routes/routes.js
@@ -1,0 +1,8 @@
+export default {
+  signUpPage: '/sign-up',
+  signInPage: '/sign-in',
+  restorePage: '/restore',
+  doctorPatientsPage: '/doctor-patients',
+  userAppointmentsPage: '/user-appointments',
+  createAppointmentPage: '/create-appointment',
+};
